+++ lib/roda/plugins/data_grid.rb
@@ -156,7 +156,7 @@ class Roda
 
         def search_presenter(id, dmc, params, for_rerun)
           OpenStruct.new(rpt: dmc.report,
-                         qps: dmc.report.ordered_query_parameter_definitions,
+                         qps: dmc.parameter_list,
                          rpt_id: id,
                          load_params: (params[:back] && params[:back] == 'y'),
                          rerun: for_rerun)